+++ app/src/main/java/com/devband/tronwalletforandroid/tron/repository/FileRepository.java
@@ -22,7 +22,7 @@ public class FileRepository implements AccountRepository {
     }
 
     @Override
-    public Maybe<AccountModel> loadAccount(int index) {
+    public Maybe<AccountModel> loadAccount(long index) {
         return null;
     }
